Changed to pass by reference
Basically, the CopyInstr function takes a MemInstr "target" by
reference, copies "this" into "target". This allows the MemInstr created
in the N-1 stack frame (which is then passed into NxtPkg by reference)
to be populated with all the data from the freshly constructed MemInstr
which will die as the NxtPkg function returns.